+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 This release extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 
 #### Features supported
+- Artifact Bundle
+- Deployment Group
 - Deployment Plan
 - Enclosure
 - Enclosure Group